Addressing comments + refactor
