+++ release/release_tests.yaml
@@ -3319,6 +3319,31 @@
 # Core Chaos tests
 ##################
 
+- name: chaos_many_tasks_baseline
+  python: "3.10"
+  group: core-nightly-test
+  working_dir: nightly_tests
+
+  frequency: nightly
+  team: core
+  cluster:
+    byod: {}
+    cluster_compute: chaos_test/compute_template.yaml
+
+  run:
+    timeout: 3600
+    wait_for_nodes:
+      num_nodes: 10
+    script: python chaos_test/test_chaos.py --workload=tasks --disable-resource-killer
+
+  variations:
+    - __suffix__: aws
+    - __suffix__: gce
+      env: gce
+      frequency: manual
+      cluster:
+        cluster_compute: chaos_test/compute_template_gce.yaml
+
 - name: chaos_many_tasks_kill_raylet
   python: "3.10"
   group: core-nightly-test
@@ -3335,7 +3360,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start
-    script: python chaos_test/test_chaos_basic.py --workload=tasks
+    script: python chaos_test/test_chaos.py --workload=tasks
 
   variations:
     - __suffix__: aws
@@ -3361,7 +3386,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start --chaos TerminateEC2Instance
-    script: python chaos_test/test_chaos_basic.py --workload=tasks
+    script: python chaos_test/test_chaos.py --workload=tasks
 
   variations:
     - __suffix__: aws
@@ -3387,7 +3412,32 @@
     wait_for_nodes:
       num_nodes: 10
     script: >
-      python simulate_cross_az_network_failure.py --network-failure-interval 180 --network-failure-duration 5 --command python chaos_test/test_chaos_basic.py --workload=tasks
+      python simulate_cross_az_network_failure.py --network-failure-interval 180 --network-failure-duration 5 --command python chaos_test/test_chaos.py --workload=tasks --disable-resource-killer
+
+  variations:
+    - __suffix__: aws
+    - __suffix__: gce
+      env: gce
+      frequency: manual
+      cluster:
+        cluster_compute: chaos_test/compute_template_gce.yaml
+
+- name: chaos_many_actors_baseline
+  python: "3.10"
+  group: core-nightly-test
+  working_dir: nightly_tests
+
+  frequency: nightly
+  team: core
+  cluster:
+    byod: {}
+    cluster_compute: chaos_test/compute_template.yaml
+
+  run:
+    timeout: 4200
+    wait_for_nodes:
+      num_nodes: 10
+    script: python chaos_test/test_chaos.py --workload=actors --disable-resource-killer
 
   variations:
     - __suffix__: aws
@@ -3413,7 +3463,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start
-    script: python chaos_test/test_chaos_basic.py --workload=actors
+    script: python chaos_test/test_chaos.py --workload=actors
 
   variations:
     - __suffix__: aws
@@ -3439,7 +3489,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start --chaos TerminateEC2Instance
-    script: python chaos_test/test_chaos_basic.py --workload=actors
+    script: python chaos_test/test_chaos.py --workload=actors
 
   variations:
     - __suffix__: aws
@@ -3465,7 +3515,32 @@
     wait_for_nodes:
       num_nodes: 10
     script: >
-      python simulate_cross_az_network_failure.py --network-failure-interval 180 --network-failure-duration 5 --command python chaos_test/test_chaos_basic.py --workload=actors
+      python simulate_cross_az_network_failure.py --network-failure-interval 180 --network-failure-duration 5 --command python chaos_test/test_chaos.py --workload=actors --disable-resource-killer
+
+  variations:
+    - __suffix__: aws
+    - __suffix__: gce
+      env: gce
+      frequency: manual
+      cluster:
+        cluster_compute: chaos_test/compute_template_gce.yaml
+
+- name: chaos_streaming_generator_baseline
+  python: "3.10"
+  group: core-nightly-test
+  working_dir: nightly_tests
+
+  frequency: nightly
+  team: core
+  cluster:
+    byod: {}
+    cluster_compute: chaos_test/compute_template.yaml
+
+  run:
+    timeout: 3600
+    wait_for_nodes:
+      num_nodes: 10
+    script: python chaos_test/test_chaos.py --workload=streaming --disable-resource-killer
 
   variations:
     - __suffix__: aws
@@ -3491,7 +3566,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start --max-to-kill 10
-    script: python chaos_test/test_chaos_basic.py --workload=streaming
+    script: python chaos_test/test_chaos.py --workload=streaming
 
   variations:
     - __suffix__: aws
@@ -3517,7 +3592,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start --chaos TerminateEC2Instance --max-to-kill 10
-    script: python chaos_test/test_chaos_basic.py --workload=streaming
+    script: python chaos_test/test_chaos.py --workload=streaming
 
   variations:
     - __suffix__: aws
@@ -3548,7 +3623,32 @@
     wait_for_nodes:
       num_nodes: 10
     script: >
-      python simulate_cross_az_network_failure.py --network-failure-interval 120 --network-failure-duration 5 --command python chaos_test/test_chaos_basic.py --workload=streaming
+      python simulate_cross_az_network_failure.py --network-failure-interval 120 --network-failure-duration 5 --command python chaos_test/test_chaos.py --workload=streaming --disable-resource-killer
+
+  variations:
+    - __suffix__: aws
+    - __suffix__: gce
+      env: gce
+      frequency: manual
+      cluster:
+        cluster_compute: chaos_test/compute_template_gce.yaml
+
+- name: chaos_object_ref_borrowing_baseline
+  python: "3.10"
+  group: core-nightly-test
+  working_dir: nightly_tests
+
+  frequency: nightly
+  team: core
+  cluster:
+    byod: {}
+    cluster_compute: chaos_test/compute_template.yaml
+
+  run:
+    timeout: 3600
+    wait_for_nodes:
+      num_nodes: 10
+    script: python chaos_test/test_chaos.py --workload=borrowing --disable-resource-killer
 
   variations:
     - __suffix__: aws
@@ -3574,7 +3674,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start
-    script: python chaos_test/test_chaos_basic.py --workload=borrowing
+    script: python chaos_test/test_chaos.py --workload=borrowing
 
   variations:
     - __suffix__: aws
@@ -3600,7 +3700,7 @@
     wait_for_nodes:
       num_nodes: 10
     prepare: python setup_chaos.py --no-start --chaos TerminateEC2Instance
-    script: python chaos_test/test_chaos_basic.py --workload=borrowing
+    script: python chaos_test/test_chaos.py --workload=borrowing
 
   variations:
     - __suffix__: aws
@@ -3631,7 +3731,7 @@
     wait_for_nodes:
       num_nodes: 10
     script: >
-      python simulate_cross_az_network_failure.py --network-failure-interval 120 --network-failure-duration 5 --command python chaos_test/test_chaos_basic.py --workload=borrowing
+      python simulate_cross_az_network_failure.py --network-failure-interval 120 --network-failure-duration 5 --command python chaos_test/test_chaos.py --workload=borrowing --disable-resource-killer
 
   variations:
     - __suffix__: aws
